Refactor clipboard helper for comments and share popup
