+++ app/javascript/comments.js
@@ -1,3 +1,5 @@
+import { copyTextToClipboard } from './utils/clipboard';
+
 if (!window.commentsInitialized) {
     window.commentsInitialized = true;
 
@@ -86,41 +88,6 @@ if (!window.commentsInitialized) {
         var form = document.getElementById('new-comment-form');
 
         if (!popup || !list || !form) { return; }
-
-        function fallbackCopyText(text) {
-            return new Promise(function(resolve, reject) {
-                var temp = document.createElement('textarea');
-                temp.value = text;
-                temp.setAttribute('readonly', '');
-                temp.style.position = 'fixed';
-                temp.style.opacity = '0';
-                temp.style.pointerEvents = 'none';
-                document.body.appendChild(temp);
-                var succeeded = false;
-                try {
-                    temp.focus();
-                    temp.select();
-                    succeeded = document.execCommand('copy');
-                } catch (err) {
-                    succeeded = false;
-                }
-                document.body.removeChild(temp);
-                if (succeeded) {
-                    resolve();
-                } else {
-                    reject();
-                }
-            });
-        }
-
-        function copyTextToClipboard(text) {
-            if (navigator.clipboard && navigator.clipboard.writeText) {
-                return navigator.clipboard.writeText(text).catch(function() {
-                    return fallbackCopyText(text);
-                });
-            }
-            return fallbackCopyText(text);
-        }
 
         function showCopyFeedback(commentElement, message) {
             if (!commentElement || !message) return;
